+++ packages/sources/bitex/src/endpoint/crypto.ts
@@ -4,9 +4,23 @@ import { ExecuteWithConfig, Config, InputParameters } from '@chainlink/types'
 export const supportedEndpoints = ['crypto', 'tickers']
 
 export const inputParameters: InputParameters = {
-  base: ['base', 'from', 'coin'],
-  quote: ['quote', 'to', 'market'],
-  resultPath: false,
+  base: {
+    aliases: ['from', 'coin'],
+    description: 'The symbol of the currency to query',
+    required: true,
+    type: 'string',
+  },
+  quote: {
+    aliases: ['to', 'market'],
+    description: 'The symbol of the currency to convert to',
+    required: true,
+    type: 'string',
+  },
+  resultPath: {
+    description: 'The object path to access the value that will be returned as the result',
+    default: 'vwap',
+    type: 'string',
+  },
 }
 
 interface ResponseSchema {