+++ packages/sources/coinbase/src/endpoint/crypto.ts
@@ -4,8 +4,22 @@ import { ExecuteWithConfig, Config, InputParameters } from '@chainlink/types'
 export const supportedEndpoints = ['crypto', 'price']
 
 export const inputParameters: InputParameters = {
-  symbol: ['base', 'from', 'coin', 'sym', 'symbol'],
-  convert: ['quote', 'to', 'market', 'convert'],
+  symbol: {
+    aliases: ['base', 'from', 'coin', 'sym'],
+    type: 'string',
+    required: true,
+    description: 'The symbol of the currency to query',
+  },
+  convert: {
+    aliases: ['quote', 'to', 'market'],
+    type: 'string',
+    required: true,
+    description: 'The symbol of the currency to convert to',
+  },
+}
+
+export interface ResponseSchema {
+  data: { base: string; currency: string; amount: string }
 }
 
 export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {
@@ -29,8 +43,8 @@ export const execute: ExecuteWithConfig<Config> = async (request, _, config) => 
     params,
   }
 
-  const response = await Requester.request(options)
-  response.data.result = Requester.validateResultNumber(response.data, ['data', 'amount'])
+  const response = await Requester.request<ResponseSchema>(options)
+  const result = Requester.validateResultNumber(response.data, ['data', 'amount'])
 
-  return Requester.success(jobRunID, response, config.verbose)
+  return Requester.success(jobRunID, Requester.withResult(response, result), config.verbose)
 }