+++ packages/sources/etherchain/src/endpoint/gasprice.ts
@@ -3,10 +3,14 @@ import { ExecuteWithConfig, Config, InputParameters } from '@chainlink/types'
 
 export const supportedEndpoints = ['gasprice']
 
-const customError = (data: any) => data.Response === 'Error'
-
 export const inputParameters: InputParameters = {
-  speed: false,
+  speed: {
+    required: false,
+    type: 'string',
+    description: 'The desired speed',
+    options: ['safeLow', 'standard', 'fast', 'fastest'],
+    default: 'standard',
+  },
 }
 
 interface ResponseSchema {
@@ -21,7 +25,7 @@ export const execute: ExecuteWithConfig<Config> = async (request, _, config) => 
   if (validator.error) throw validator.error
 
   const jobRunID = validator.validated.id
-  const speed = validator.validated.data.speed || 'standard'
+  const speed = validator.validated.data.speed
   const url = `/api/gasPriceOracle`
 
   const options = {
@@ -29,7 +33,7 @@ export const execute: ExecuteWithConfig<Config> = async (request, _, config) => 
     url,
   }
 
-  const response = await Requester.request<ResponseSchema>(options, customError)
+  const response = await Requester.request<ResponseSchema>(options)
   const result = Requester.validateResultNumber(response.data, [speed]) * 1e9
 
   return Requester.success(jobRunID, Requester.withResult(response, result), config.verbose)