+++ packages/sources/expert-car-broker/src/endpoint/feed.ts
@@ -4,8 +4,20 @@ import { Config, ExecuteWithConfig, InputParameters } from '@chainlink/types'
 export const supportedEndpoints = ['feed']
 
 export const inputParameters: InputParameters = {
-  product: true,
-  feedId: true,
+  product: {
+    required: true,
+    description: 'The product to query',
+    type: 'string',
+  },
+  feedId: {
+    required: true,
+    description: 'The feed ID to use',
+    type: 'number',
+  },
+}
+
+export interface ResponseSchema {
+  value: number
 }
 
 export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {
@@ -23,8 +35,8 @@ export const execute: ExecuteWithConfig<Config> = async (request, _, config) => 
 
   const options = { ...config.api, params, url }
 
-  const response = await Requester.request(options)
-  response.data.result = Requester.validateResultNumber(response.data, ['value'])
+  const response = await Requester.request<ResponseSchema>(options)
+  const result = Requester.validateResultNumber(response.data, ['value'])
 
-  return Requester.success(jobRunID, response, config.verbose)
+  return Requester.success(jobRunID, Requester.withResult(response, result), config.verbose)
 }