+++ packages/sources/twelvedata/src/endpoint/closing.ts
@@ -7,7 +7,20 @@ export const supportedEndpoints = ['closing', 'eod']
 const customError = (data: any) => data.Response === 'Error'
 
 export const inputParameters: InputParameters = {
-  base: ['base', 'from', 'coin', 'market', 'symbol'],
+  base: {
+    aliases: ['from', 'coin', 'market', 'symbol'],
+    required: true,
+    description: 'The symbol of the currency to query',
+    type: 'string',
+  },
+}
+
+interface ResponseSchema {
+  symbol: string
+  exchange: string
+  currency: string
+  datetime: string
+  close: string
 }
 
 export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {
@@ -29,8 +42,9 @@ export const execute: ExecuteWithConfig<Config> = async (request, _, config) => 
     url,
   }
 
-  const response = await Requester.request(options, customError)
-  response.data.result = Requester.validateResultNumber(response.data, ['close'])
+  const response = await Requester.request<ResponseSchema>(options, customError)
 
-  return Requester.success(jobRunID, response, config.verbose)
+  const result = Requester.validateResultNumber(response.data, ['close'])
+
+  return Requester.success(jobRunID, Requester.withResult(response, result), config.verbose)
 }