+++ packages/composites/google-weather/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @chainlink/google-weather-adapter
 
+## 1.0.16
+
+### Patch Changes
+
+- Updated dependencies [f272a595]
+- Updated dependencies [1d55bbde]
+- Updated dependencies [6feacbc2]
+  - @chainlink/ea-bootstrap@1.6.0
+  - @chainlink/google-bigquery-adapter@1.1.0
+
+## 1.0.15
+
+### Patch Changes
+
+- Updated dependencies [9b3cd511]
+  - @chainlink/ea-bootstrap@1.5.0
+  - @chainlink/google-bigquery-adapter@1.0.14
+
 ## 1.0.14
 
 ### Patch Changes