+++ packages/composites/reference-transform/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @chainlink/reference-transform-adapter
 
+## 1.0.28
+
+### Patch Changes
+
+- Updated dependencies [f272a595]
+- Updated dependencies [1d55bbde]
+  - @chainlink/ea-bootstrap@1.6.0
+  - @chainlink/ea@1.0.28
+  - @chainlink/ea-reference-data-reader@1.0.15
+
+## 1.0.27
+
+### Patch Changes
+
+- Updated dependencies [9b3cd511]
+  - @chainlink/ea-bootstrap@1.5.0
+  - @chainlink/ea@1.0.27
+  - @chainlink/ea-reference-data-reader@1.0.14
+
+## 1.0.26
+
+### Patch Changes
+
+- @chainlink/ea@1.0.26
+
 ## 1.0.25
 
 ### Patch Changes