+++ packages/core/bootstrap/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @chainlink/ea-bootstrap
 
+## 1.6.0
+
+### Minor Changes
+
+- f272a595: Implement cache fallback method
+- 1d55bbde: Change data provider error surfacing, return 200 and use new providerStatusCode field
+
+## 1.5.0
+
+### Minor Changes
+
+- 9b3cd511: limit number of times ws message handled
+
 ## 1.4.0
 
 ### Minor Changes