+++ packages/sources/binance/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @chainlink/binance-adapter
 
+## 1.2.0
+
+### Minor Changes
+
+- bea9f065: input parameters changed
+
+### Patch Changes
+
+- Updated dependencies [f272a595]
+- Updated dependencies [1d55bbde]
+  - @chainlink/ea-bootstrap@1.6.0
+  - @chainlink/ea-test-helpers@1.0.1
+
+## 1.1.7
+
+### Patch Changes
+
+- Updated dependencies [9b3cd511]
+  - @chainlink/ea-bootstrap@1.5.0
+  - @chainlink/ea-test-helpers@1.0.1
+
+## 1.1.6
+
+### Patch Changes
+
+- 8ee3290d: added response types for crypto endpoint, removed customError function
+
 ## 1.1.5
 
 ### Patch Changes