+++ packages/sources/coinmetrics/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @chainlink/coinmetrics-adapter
 
+## 1.2.0
+
+### Minor Changes
+
+- 6b7f9d56: input parameters changed
+
+### Patch Changes
+
+- Updated dependencies [f272a595]
+- Updated dependencies [1d55bbde]
+  - @chainlink/ea-bootstrap@1.6.0
+  - @chainlink/ea-test-helpers@1.0.1
+
+## 1.1.14
+
+### Patch Changes
+
+- Updated dependencies [9b3cd511]
+  - @chainlink/ea-bootstrap@1.5.0
+  - @chainlink/ea-test-helpers@1.0.1
+
 ## 1.1.13
 
 ### Patch Changes