+++ packages/sources/tiingo/CHANGELOG.md
@@ -1,5 +1,31 @@
 ## v1.1.0
 
+## 1.6.0
+
+### Minor Changes
+
+- 427e27da: only create default config once for tiingo ws
+
+### Patch Changes
+
+- f26b3896: added response types
+- Updated dependencies [f272a595]
+- Updated dependencies [1d55bbde]
+  - @chainlink/ea-bootstrap@1.6.0
+  - @chainlink/ea-test-helpers@1.0.1
+
+## 1.5.0
+
+### Minor Changes
+
+- 9b3cd511: limit number of times ws message handled
+
+### Patch Changes
+
+- Updated dependencies [9b3cd511]
+  - @chainlink/ea-bootstrap@1.5.0
+  - @chainlink/ea-test-helpers@1.0.1
+
 ## 1.4.0
 
 ### Minor Changes